pyproject.toml
[tool.poetry]
name = "nightingale-backend"
version = "0.0.0.post7.dev0+eaeb93c"
description = ""
authors = ["GLEF1X <glebgar567@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.96.0"
uvicorn = { version = "^0.22.0", extras = ["standard"] }
python-dateutil = "^2.8.2"
python-dotenv = "^1.0.0"
httpx = "^0.24.0"
itsdangerous = "^2.1.2"
fasteners = "^0.18"
orjson = "^3.9.0"
email-validator = "^2.0.0.post2"
beanie = "^1.19.0"
starlette-context = "^0.3.6"
pyjwt = { version = "^2.7.0", extras = ["crypto"]}
structlog = "^23.1.0"
colorama = "^0.4.6"
pyhumps = "^3.8.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.2.0"
ruff = "^0.0.265"
isort = "^5.12.0"
pytest = "^7.3.1"
pytest-anyio = "^0.0.0"
motor-stubs = "^1.7.1"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = [
    "tests",
]

[tool.ruff]
fix = true
unfixable = [
    "ERA", # do not autoremove commented out code
]
target-version = "py310"
line-length = 100
extend-select = [
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "ERA", # flake8-eradicate/eradicate
    "I",   # isort
    "N",   # pep8-naming
    "PIE", # flake8-pie
    "PGH", # pygrep
    "RUF", # ruff checks
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "TID", # flake8-tidy-imports
    "UP",  # pyupgrade
]

[tool.ruff.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.isort]
force-single-line = true
lines-between-types = 1
lines-after-imports = 2
required-imports = ["from __future__ import annotations"]

[tool.mypy]
plugins = "pydantic.mypy"
python_version = "3.11"
show_error_codes = true
show_error_context = true
pretty = true
ignore_missing_imports = false
warn_unused_configs = true
disallow_subclassing_any = true
disallow_any_generics = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
follow_imports_for_stubs = true
namespace_packages = true
show_absolute_path = true

[[tool.mypy.overrides]]
module = [
    "aiofiles",
    "async_lru",
    "uvloop",
    "redis.*",
    "babel.*",
]
ignore_missing_imports = true
disallow_untyped_defs = true

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "pep440"

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"
